Change debug launch to use socket-listen instead of socket-connect (#247)
Launching the dev_appserver in debug mode often fails due to timing
issues.  Although we have a 500ms wait, it seems insufficient
to ensure the remote JVM has gotten to a debuggable state and
so the debug connection fails.  We instead use the socket-listen
launch, such that the remote JVM connects to the debugger.