Code Drop of the Dataflow Plugin for Eclipse Core (#1508)
This is part of a sequence of commits to move the Dataflow Eclipse
Plugin to the Cloud Tools for Eclipse plugin. This is primarily a port
of the Dataflow Core Plugin, with minimal modifications made as
required to support proper dependency resolution, style migrations,
and other miscellaneous fixes.

* Port Dataflow Preferences
* Port Dataflow Property Testers
* Port Dataflow PipelineOptions Classes
* Port Dataflow Project Management classes
* Port Dataflow Core Plugin utils
* Port Dataflow Launching

Additionally the following changes are made to the Dataflow plugin:
Deprecate "trackDataflowDependency". LATEST and RELEASE are deprecated
Replace Set in DataflowPreferences with an ImmutableSet